+++ calendar-bundle/src/Resources/contao/classes/Calendar.php
@@ -342,7 +342,7 @@ class Calendar extends \Frontend
 						}
 
 						// The target page is exempt from the sitemap (see #6418)
-						if ($objParent->sitemap == 'map_never')
+						if ($objParent->robots == 'noindex,nofollow')
 						{
 							continue;
 						}