+++ faq-bundle/src/Resources/contao/modules/ModuleFaq.php
@@ -84,7 +84,7 @@ class ModuleFaq extends \Frontend
 						}
 
 						// The target page is exempt from the sitemap (see #6418)
-						if ($objParent->sitemap == 'map_never')
+						if ($objParent->robots == 'noindex,nofollow')
 						{
 							continue;
 						}