+++ newsletter-bundle/src/Resources/contao/classes/Newsletter.php
@@ -989,7 +989,7 @@ class Newsletter extends \Backend
 						}
 
 						// The target page is exempt from the sitemap (see #6418)
-						if ($objParent->sitemap == 'map_never')
+						if ($objParent->robots == 'noindex,nofollow')
 						{
 							continue;
 						}